Downloading from remote data source

Two epynomous methods are provided for downloading free data, from both Yahoo Finance (yahoo) and the Federal Reserve St. Louis (fred).

These methods take a string argument that represents the name of the desired data set to be downloaded.

These methods have default arguments. These defaults only apply when no argument is passed, otherwise the user is left with determining the correct name for the data of interest.

Yahoo

For the yahoo() method, the default data is the S&P 500 end-of-day daily prices. That symbol is named ^GSPC.

MarketData.yahoo — Function

yahoo(symbol::AbstractString, opt::YahooOpt = YahooOpt())::TimeArray
yahoo(symbol::Symbol, opt::YahooOpt = YahooOpt())::TimeArray

This is a wrapper for downloading historical stock prices from Yahoo Finance.

The yahoo method takes a stock name in the form of a string and returns a TimeSeries.TimeArray corresponding to the Yahoo Finance ticker. With no argument, the default historical time series is the S&P 500.

Examples

AAPL = yahoo(:AAPL)
SPX = yahoo("^GSPC")
NQ = yahoo("^IXIC")

julia> start = DateTime(2018, 1, 1)
2018-01-01T00:00:00

julia> yahoo(:AAPL, YahooOpt(period1 = start))
655×6 TimeArray{Float64,2,Date,Array{Float64,2}} 2018-01-02 to 2020-08-07
...

References

https://finance.yahoo.com

See Also

fred() which accesses the St. Louis Federal Reserve financial and economic data sets.

MarketData.YahooOpt — Type

struct YahooOpt <: AbstractQueryOpt
  period1  # the start time
  period2  # the end time
  interval # "1d", "1wk" or "1mo"
  events   # currently only `:history` supported
end

The Yahoo Finance HTTP API query object.

Examples

julia> t = Dates.now()
2020-08-09T01:38:04.735

julia> YahooOpt(period1 = t - Year(2), period2 = t)
YahooOpt{DateTime} with 4 entries:
  :period1  => 1533778685
  :period2  => 1596937085
  :interval => "1d"
  :events   => :history

FRED

The default data for the fred() method is CPIAUCNS, which represents the Consumer Price Index for All Urban Consumers data.